Enable \watch command to use fractional seconds (ala psql) and to support standard Go durations.
psql's \watch command allows fractional seconds, ie:

  \watch 0.11

This commit enables the above, as well as standard Go duration parsing
on passed commands:

  \watch 0.11
  \watch 0.11s
  \watch 1h33m

(when a duration modifier is not specified, defaults to interpreting it
as seconds, ala standard Unix commands and psql)

Additionally, slightly modifies the \watch output to add an additional
empty line between the time output and the table output, to be closer to
psql's actual output. Further, cleans up the watch implementation and
propagated execution options to usql's handler, changing the name of the
metacmd.Result -> metacmd.Option to be more clear to use/purpose.